Changed getHash to a local function of QRCode and removed unnecesary Hash class
