--- QRiffic Application/app/src/test/java/com/example/qriffic/HashTest.java
@@ -1,22 +1,0 @@
-package com.example.qriffic;
-
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
-
-
-
-/**
- * Test suite for Hash class
- */
-public class HashTest {
-
-    @Test
-    void testGetters() {
-
-        // SHA256 output generated by https://emn178.github.io/online-tools/sha256.html
-        assertEquals("ca978112ca1bbdcafac231b39a23dc4da786eff8147c4e72b9807785afee48bb",
-                Hash.getHash("a"));
-        assertNotEquals("ca978112ca1bbdcafac231b39a23dc4da888eff8147c4e72b9807785afee48bb",
-                Hash.getHash("a"));
-    }
-}
